removed code that conflicted with previous change
