+++ run/server.properties
@@ -1,5 +1,5 @@
 #Minecraft server properties
-#Sun Jan 29 13:33:02 CST 2023
+#Sun Jan 29 14:19:27 CST 2023
 enable-jmx-monitoring=false
 rcon.port=25575
 level-seed=
