+++ run/world/stats/0e35bb95-948e-4baf-813c-5e285bfda781.json
@@ -1,1 +1,1 @@
-{"stats":{"minecraft:picked_up":{"minecraft:rotten_flesh":19},"minecraft:custom":{"minecraft:jump":355,"minecraft:time_since_rest":1629,"minecraft:crouch_one_cm":12,"minecraft:damage_taken":60001300,"minecraft:walk_one_cm":22359,"minecraft:leave_game":2,"minecraft:sprint_one_cm":20900,"minecraft:play_time":9539,"minecraft:time_since_death":1628,"minecraft:walk_on_water_one_cm":3356,"minecraft:deaths":14,"minecraft:sneak_time":13,"minecraft:total_world_time":9539,"minecraft:walk_under_water_one_cm":3865,"minecraft:swim_one_cm":10595,"minecraft:fall_one_cm":17433,"minecraft:fly_one_cm":154932}},"DataVersion":2975}
+{"stats":{"minecraft:custom":{"minecraft:jump":462,"minecraft:time_since_rest":1110,"minecraft:damage_taken":70001550,"minecraft:walk_one_cm":23925,"minecraft:crouch_one_cm":12,"minecraft:leave_game":4,"minecraft:sprint_one_cm":29833,"minecraft:play_time":12018,"minecraft:time_since_death":1110,"minecraft:walk_on_water_one_cm":3761,"minecraft:deaths":16,"minecraft:sneak_time":13,"minecraft:total_world_time":12018,"minecraft:walk_under_water_one_cm":4266,"minecraft:swim_one_cm":11552,"minecraft:fall_one_cm":23150,"minecraft:fly_one_cm":221468},"minecraft:picked_up":{"minecraft:rotten_flesh":20}},"DataVersion":2975}
