+++ src/main/java/com/geekazodium/unnamedminecraftrpg/entities/treasure/LootChestHandler.java
@@ -1,10 +1,23 @@
 package com.geekazodium.unnamedminecraftrpg.entities.treasure;
 
+import com.destroystokyo.paper.event.player.PlayerConnectionCloseEvent;
+import com.geekazodium.unnamedminecraftrpg.GameTickHandler;
+import com.geekazodium.unnamedminecraftrpg.Main;
+import com.geekazodium.unnamedminecraftrpg.items.weapons.WeaponItemHandler;
+import com.geekazodium.unnamedminecraftrpg.players.PlayerHandler;
+import org.bukkit.Bukkit;
 import org.bukkit.Effect;
 import org.bukkit.Material;
 import org.bukkit.Particle;
 import org.bukkit.block.data.BlockData;
+import org.bukkit.craftbukkit.v1_18_R2.entity.CraftPlayer;
+import org.bukkit.craftbukkit.v1_18_R2.inventory.CraftInventoryPlayer;
 import org.bukkit.entity.Player;
+import org.bukkit.event.EventHandler;
+import org.bukkit.event.HandlerList;
+import org.bukkit.event.Listener;
+import org.bukkit.event.inventory.InventoryCloseEvent;
+import org.bukkit.inventory.Inventory;
 
 import java.util.Random;
 
@@ -43,6 +56,31 @@ public abstract class LootChestHandler implements TreasureEntity{
         timeSinceLastOpened = -timeUntilRespawn;
     }
 
+    protected void openChestInventory(Player player, Inventory inventory){
+        PlayerHandler handler = GameTickHandler.getPlayerHandler(player);
+        handler.setAllowSendInventoryUpdatePacket(true);
+        PlayerHandler.sendInventoryUpdatePacket(((CraftPlayer) player).getHandle().networkManager, ((CraftInventoryPlayer) player.getInventory()).getInventory());
+        PlayerCloseInventoryListener closeInventoryListener = new PlayerCloseInventoryListener(handler, inventory);
+        Bukkit.getPluginManager().registerEvents(closeInventoryListener, Main.getInstance());
+        player.openInventory(inventory);
+    }
+
+    private record PlayerCloseInventoryListener(PlayerHandler player, Inventory inventory) implements Listener {
+        @EventHandler
+        public void onPlayerDisconnect(PlayerConnectionCloseEvent event){
+            if(event.getPlayerUniqueId()!=player.getPlayer().getUniqueId())return;
+            HandlerList.unregisterAll(this);
+            player.setAllowSendInventoryUpdatePacket(false);
+        }
+
+        @EventHandler
+        public void onPlayerCloseInventory(InventoryCloseEvent event){
+            if(event.getPlayer()!=player.getPlayer())return;
+            if(event.getInventory()!=inventory)return;
+            player.sendVisualOnlyWeaponHotbar(player.getPlayer().getInventory(), (WeaponItemHandler) player.getActiveItemHandler());
+            player.setAllowSendInventoryUpdatePacket(false);
+        }
+    }
     @Override
     public boolean canInteract(Player player) {
         return isPlaced;
